Update README.md
HoreenX: The 1st Open Source AiWare for #MusicLearning and #SoundLearning. You are not just listening to the Musics and the Sounds, NOW! You are running an ACSGPT Neural Computer Program. With the Ai-Program of @jawolusy, Human Machines can now Learn from the Music and the Sounds. This Ai-Framework is usable for Audio Passkey Services, HaddJay Neural Banking Services, and to develop several other MultiLingual Artificial Intelligence Applications. The Nigerian #CelebCoded Neural Computer is the default template of the HoreenX Ai-App, it can always be edited and re-used to develop other National Ai-Programs, such as; the American #CelebCoded Neural Computer, the British #CelebCoded Neural Computer, the Kenyan #CelebCoded Neural Computer, the #64HymnsLLM Ai-App, and so on.